Build using gradle, fix runner
