+++ agent/build.gradle
@@ -1,0 +1,28 @@
+plugins {
+    id 'org.jetbrains.intellij' version '1.1.3'
+    id 'java'
+}
+
+repositories {
+    mavenCentral()
+}
+
+intellij {
+    version 'LATEST-EAP-SNAPSHOT'
+    downloadSources = true
+}
+
+jar {
+    manifest {
+        attributes("Main-Class": "unit.modified.com.intellij.rt.coverage.main.CoveragePremain",
+                "Manifest-Version": "1.0",
+                "Premain-Class": "unit.modified.com.intellij.rt.coverage.main.CoveragePremain",
+                "Agent-Class": " unit.modified.com.intellij.rt.coverage.main.CoveragePremain",
+                "Can-Redefine-Classes": "true",
+                "Can-Retransform-Classes": "true",
+        )
+    }
+
+    from(zipTree("${runIde.ideDir.get()}/lib/intellij-coverage.jar"))
+}
+
