--- agent/src/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: unit.modified.com.intellij.rt.coverage.main.CoveragePremain
-
