+++ plugin/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id 'org.jetbrains.intellij' version '1.1.3'
+    id 'java'
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation project(':agent')
+}
+
+// See https://github.com/JetBrains/gradle-intellij-plugin/
+intellij {
+    version 'LATEST-EAP-SNAPSHOT'
+    downloadSources = true
+    plugins = ['com.intellij.java', 'JUnit', 'TestNG-J']
+}
+test {
+    useJUnitPlatform()
+}
