+++ plugin/src/main/java/gabor/unittest/history/action/ResetAction.java
