+++ plugin/src/main/java/gabor/unittest/history/debug/DebugResources.java
@@ -1,0 +1,6 @@
+package gabor.unittest.history.debug;
+
+public class DebugResources {
+    //these should be changed when debug module is ported to another plugin
+    public static final String WITH_HISTORY_DEBUGGER = "Run with Unit Test History Coverage";
+}
