+++ plugin/src/main/java/gabor/unittest/history/helper/ConfigType.java
