--- unit-test-history-runner/resources/META-INF/plugin.xml
@@ -1,80 +1,0 @@
-<idea-plugin>
-    <id>Unit-Test-Coverage-Viewer</id>
-    <name>Unit Test Coverage History Runner</name>
-    <version>1.17</version>
-    <vendor
-            email="csabag.laszlo@gmail.com"
-            url="https://github.com/csabagabor">
-        Csaba Gabor
-    </vendor>
-
-    <category>Unit testing</category>
-
-    <description><![CDATA[
-
-   <br>
-<b>Important: it only works with JDK 8 or higher, supports JUnit 3,4,5 and TestNG</b>
-<br>
-This plugin lets you <b>click on a method in your project and automatically run all the unit tests which cover that method.</b>
-<br>
-<br>
-<b>Manual (Help):</b>
-<p><b>There are two steps to use the plugin:</b>
-<ul>
-<li>1. First you have to run as many unit tests as you can <b>with the new Unit Test Runner</b>(its icon has 1 red + 1 green triangle and a yellow arrow on top of the triangles).
- The more tests you run with it, the more coverage information you will have.</li>
- <li>2. To run all the unit tests which cover a particular method in your project, just right click on a method
-and then choose -> <b>Run Unit Tests Which Cover This Method...(this is not shown for methods without coverage info)</b> Unit tests which cover that method will be automatically started.</li>
-</ul>
-<br>
-Also to <b>reset coverage information</b> (in case unexpected tests are run), you can go to Tools -> Reset Coverage Info...
-<br>
-<br>
-<b>Limitations:</b>
-When you encounter any test for which coverage information is not included, contact me.
-For example the plugin doesn't work when using empty test methods with the Scenario annotation like <code>@Scenario("integration_tests")</code>
-<br>
-
-Credits:<br><br>
-IntelliJ IDEA Code Coverage Agent: https://github.com/JetBrains/intellij-coverage
-    ]]></description>
-
-    <change-notes><![CDATA[
-        <p>1.17 (13-sept-2020)</p>
-        <ul>
-            <li>Initial version with support for JUnit 3,4,5 and TestNG </li>
-        </ul>
-    ]]>
-    </change-notes>
-
-    <!-- please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description -->
-    <idea-version since-build="193"/>
-
-    <!-- please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/plugin_compatibility.html
-         on how to target different products -->
-
-    <depends>com.intellij.modules.java</depends>
-    <depends>JUnit</depends>
-    <depends>TestNG-J</depends>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <projectService serviceImplementation="gabor.unittest.history.action.CoverageContext"/>
-        <executor implementation="gabor.unittest.history.debug.executor.CustomHistoryDebuggerExecutor"/>
-        <programRunner implementation="gabor.unittest.history.debug.executor.CustomHistoryDebugger" order="last"/>
-    </extensions>
-
-    <actions>
-        <action id="gabor.unittest.history.action.ResetAction"
-                class="gabor.unittest.history.action.ResetAction"
-                text="Reset Coverage Info..." description="Load breakpoint from file..."
-                icon="/icons/history/debug.svg">
-            <add-to-group group-id="ToolsMenu" relative-to-action="GenerateJavadoc" anchor="after"/>
-        </action>
-        <action id="gabor.unittest.history.action.ShowHistoryForMethodAction"
-                class="gabor.unittest.history.action.ShowHistoryForMethodAction"
-                text="Run Unit Tests Which Cover This Method..." description="Run unit tests which cover this method..."
-                icon="/icons/history/debug.svg">
-            <add-to-group group-id="EditorPopupMenu"/>
-        </action>
-    </actions>
-</idea-plugin>
